chore: cleanup function signatures (#215)
This resolves INT-4552.

More specifically, this does three things:
- [x] Adds a CI check for matching signatures in `crates/tracegen`
between Rust and CUDA. It already found one number-of-arguments-mismatch
in some sha function, one we-pass-`u32`-but-cuda-reads-`size_t` in field
expression tracegen (although cuda code instantly converted it back to
`uint32_t` so it didn't trigger), and another error of the same sort for
some other function, where a lot of parameters we passed were `u32` but
needed to be `size_t`. It also caught a lot of reverse mismatches, where
we passed `size_t` and received `u32`, which I think was not a big deal,
but I fixed them anyway. The two scripts I mostly got from chat gpt are
now in `tracegen/scripts`.
- [x] Adds a struct that holds a pointer with some size in bytes, aka a
view to a `DeviceBuffer`. In Rust this is just `struct DeviceBufferView
{ ptr: *const c_void, size: usize }`, in CUDA I decided to make two
separate structures `DeviceRawBufferConstView { uintptr_t ptr; size_t
size; }`, and `template <typename T> struct DeviceBufferConstView` that
also provides easy access (aka `operator []`). The `DeviceBufferView` in
Rust has the same alignment etc as those three CUDA structs, because 1)
it does and 2) the script above doesn't complain. A lot of places where
we just passed a pointer and a length or `num_records` now become
simpler, and we do a lot fewer
`reinterpret_cast<RecordType*>(records)[idx]`. This PR doesn't claim to
change all of these places, but probably most of them.
- [x] We now just pass trace buffers as `*mut F` instead of
`tracegen<T>(d_trace: &DeviceBuffer<T>)` or anything. On the receiving
side this is treated as `Fp* d_trace` anyway.

---------

+++ tracegen-gpu/cuda/src/mod_builder/field_expression.cu
@@ -346,7 +346,7 @@ extern "C" int _field_expression_tracegen(
     uint32_t range_checker_num_bins,
     uint32_t *d_bitwise_lookup,
     uint32_t bitwise_num_bits,
-    size_t pointer_max_bits,
+    uint32_t pointer_max_bits,
     uint32_t timestamp_max_bits,
     uint8_t *d_workspace,
     uint32_t workspace_per_thread
